Build-Week-Rule-Engine User-Interface #2 update
I decided to change the polygon background image from the main page file because of the word on the image and replace it with a linear-vector-chart image. I didn't get to make any changes on the media queries because I had to take a javascript test. I added a footer to the main page.